requirements.txt
numpy==1.24.2
pandas==1.5.3
python-dotenv==1.0.0
streamlit==1.35.0
openai==1.53.0
ipykernel==6.29.5